Added changes (WebDrioverManager, before and after) to previous HW TanyaKraskovaTest
